# Loki: simplify the FetchChunks method and optimize the experimental l2 cache a little more