--- a/pkg/storage/stores/series/series_store_utils.go
+++ b/pkg/storage/stores/series/series_store_utils.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/storage/chunk"
-	"github.com/grafana/loki/pkg/storage/config"
 	"github.com/grafana/loki/pkg/util"
 )
 
@@ -48,9 +47,8 @@ func labelNamesFromChunks(chunks []chunk.Chunk) []string {
 	return result.Strings()
 }
 
-func filterChunksByUniqueFingerprint(s config.SchemaConfig, chunks []chunk.Chunk) ([]chunk.Chunk, []string) {
+func filterChunksByUniqueFingerprint(chunks []chunk.Chunk) []chunk.Chunk {
 	filtered := make([]chunk.Chunk, 0, len(chunks))
-	keys := make([]string, 0, len(chunks))
 	uniqueFp := map[model.Fingerprint]struct{}{}
 
 	for _, chunk := range chunks {
@@ -58,15 +56,13 @@ func filterChunksByUniqueFingerprint(s config.SchemaConfig, chunks []chunk.Chunk
 			continue
 		}
 		filtered = append(filtered, chunk)
-		keys = append(keys, s.ExternalKey(chunk.ChunkRef))
 		uniqueFp[chunk.FingerprintModel()] = struct{}{}
 	}
-	return filtered, keys
+	return filtered
 }
 
-func filterChunkRefsByUniqueFingerprint(s config.SchemaConfig, chunks []logproto.ChunkRef) ([]chunk.Chunk, []string) {
+func filterChunkRefsByUniqueFingerprint(chunks []logproto.ChunkRef) []chunk.Chunk {
 	filtered := make([]chunk.Chunk, 0, len(chunks))
-	keys := make([]string, 0, len(chunks))
 	uniqueFp := map[model.Fingerprint]struct{}{}
 
 	for _, c := range chunks {
@@ -76,10 +72,9 @@ func filterChunkRefsByUniqueFingerprint(s config.SchemaConfig, chunks []logproto
 		filtered = append(filtered, chunk.Chunk{
 			ChunkRef: c,
 		})
-		keys = append(keys, s.ExternalKey(c))
 		uniqueFp[c.FingerprintModel()] = struct{}{}
 	}
-	return filtered, keys
+	return filtered
 }
 
 func uniqueStrings(cs []string) []string {