… pkg/storage/stores/series/series_store_test.go
… pkg/storage/stores/shipper/index/compactor/util_test.go