**What this PR does / why we need it**:

This started as a simple PR to optimize the new L2 chunks cache code such that we don't look up chunks in L1 that we know won't be there, however, the complexity of the FetchChunks method taking both a slice of requested Chunks as well as a slice of requested Keys made this really challenging.

Looking at how we use this method, I saw no reason for there to be a requirement to pass both a list of chunks and keys, we only ever want the Chunks and everywhere we used this we were generating the list of keys external to this function from the list of chunks being passed in.

The second change was to the implementation of how we process the cache results to look for found vs missing keys.  The current implementation required the chunks to be sorted by external key so that it could iterate through them, I tested out a different version of this which uses a map instead of sorting and iterating.

The map performance was slightly better in terms of reduced allocations, not enough to make a meaningful difference, but I think this code is easier to understand and reason about, and it removes any requirement for sorting of the passed in Chunks by External key.

Benchmarks:

Previous:
```
100 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32           3445            334642 ns/op          478074 B/op       7536 allocs/op
BenchmarkFetch-32           2947            344439 ns/op          478534 B/op       7537 allocs/op
BenchmarkFetch-32           3489            339239 ns/op          478092 B/op       7536 allocs/op
BenchmarkFetch-32           3523            340550 ns/op          478003 B/op       7536 allocs/op
BenchmarkFetch-32           2935            346810 ns/op          478590 B/op       7537 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       23.948s

1000 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32            363           3225519 ns/op         5305481 B/op      78133 allocs/op
BenchmarkFetch-32            358           3308456 ns/op         5308299 B/op      78143 allocs/op
BenchmarkFetch-32            336           3379067 ns/op         5322534 B/op      78191 allocs/op
BenchmarkFetch-32            357           3330118 ns/op         5308854 B/op      78145 allocs/op
BenchmarkFetch-32            349           3390415 ns/op         5313813 B/op      78162 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       25.607s

10,000 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32              3         334471200 ns/op        639282104 B/op   2630192 allocs/op
BenchmarkFetch-32              2         503434741 ns/op        922833480 B/op   3412973 allocs/op
BenchmarkFetch-32              1        1064668769 ns/op        1771502096 B/op  5741173 allocs/op
BenchmarkFetch-32              1        1027755910 ns/op        1771419128 B/op  5749049 allocs/op
BenchmarkFetch-32              1        1027270578 ns/op        1774711872 B/op  5752726 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       23.251s
```

Now
```
100 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32           3856            300569 ns/op          462932 B/op       6214 allocs/op
BenchmarkFetch-32           4016            305387 ns/op          462803 B/op       6214 allocs/op
BenchmarkFetch-32           3382            303078 ns/op          463201 B/op       6215 allocs/op
BenchmarkFetch-32           3319            303367 ns/op          463308 B/op       6215 allocs/op
BenchmarkFetch-32           3289            356986 ns/op          463341 B/op       6215 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       22.879s

10,000 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32            393           2641470 ns/op         5193282 B/op      65956 allocs/op
BenchmarkFetch-32            433           2684287 ns/op         5174832 B/op      65894 allocs/op
BenchmarkFetch-32            427           2778092 ns/op         5177334 B/op      65902 allocs/op
BenchmarkFetch-32            418           2714370 ns/op         5181193 B/op      65915 allocs/op
BenchmarkFetch-32            415           2822163 ns/op         5183436 B/op      65920 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       25.053s

10,000 chunks
❯ go test -bench=Fetch -count=5
goos: linuxbench=Fetch -count=5                                                                                                                                                                                                                                                         ─╯
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/chunk/fetcher
cpu: AMD Ryzen 9 5950X 16-Core Processor            
BenchmarkFetch-32              3         362233430 ns/op        638097202 B/op   2510070 allocs/op
BenchmarkFetch-32              2         611639304 ns/op        921128060 B/op   3292828 allocs/op
BenchmarkFetch-32              1        1022180867 ns/op        1770465496 B/op  5621431 allocs/op
BenchmarkFetch-32              1        1253448115 ns/op        1771278456 B/op  5630224 allocs/op
BenchmarkFetch-32              1        1189564672 ns/op        1774053064 B/op  5628113 allocs/op
PASS
ok      github.com/grafana/loki/pkg/storage/chunk/fetcher       23.960s
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
  - [ ] If the change is worth mentioning in the release notes, add `add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in `production/helm/loki/Chart.yaml` and update `production/helm/loki/CHANGELOG.md` and `production/helm/loki/README.md`. [Example PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)